Hide Alert and QuickDropPanel in drag-and-drop snopshot tests.
