--- src/e2e/puppeteer/locators/popup.ts
@@ -1,6 +1,0 @@
-import { page } from '../setup'
-
-/** Returns the popup element. */
-const popup = () => page.locator('[data-testid="popup-value"]').wait()
-
-export default popup
