--- src/e2e/puppeteer/locators/quickDropPanel.ts
@@ -1,6 +1,0 @@
-import { page } from '../setup'
-
-/** Returns the QuickDropPanel element. */
-const quickDropPanel = () => page.locator('[data-testid="quick-drop-panel"]').wait()
-
-export default quickDropPanel
